REFACTOR(fileStruct) Modified file structure of the whole program.
